Extract CreateIndexCallable
Extract CreateIndexCallable from IndexCreator making it a standalone class.
