+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/internal/query/IndexCreator.java
@@ -14,15 +14,13 @@
 
 package com.cloudant.sync.internal.query;
 
-import com.cloudant.sync.internal.android.ContentValues;
 import com.cloudant.sync.documentstore.Database;
+import com.cloudant.sync.internal.query.callables.CreateIndexCallable;
 import com.cloudant.sync.internal.query.callables.ListIndexesCallable;
 import com.cloudant.sync.query.FieldSort;
 import com.cloudant.sync.query.Index;
 import com.cloudant.sync.query.IndexType;
 import com.cloudant.sync.query.QueryException;
-import com.cloudant.sync.internal.sqlite.SQLCallable;
-import com.cloudant.sync.internal.sqlite.SQLDatabase;
 import com.cloudant.sync.internal.sqlite.SQLDatabaseFactory;
 import com.cloudant.sync.internal.sqlite.SQLDatabaseQueue;
 import com.cloudant.sync.internal.util.Misc;
@@ -30,13 +28,10 @@ import com.cloudant.sync.internal.util.Misc;
 import org.apache.commons.codec.binary.Hex;
 
 import java.nio.charset.Charset;
-import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
@@ -172,54 +167,7 @@ class IndexCreator {
         }
 
         final Index index = proposedIndex;
-        Future<Void> result = queue.submitTransaction(new SQLCallable<Void>() {
-            @Override
-            public Void call(SQLDatabase database) throws QueryException {
-
-                // Insert metadata table entries
-                for (FieldSort fieldName: fieldNamesList) {
-                    ContentValues parameters = new ContentValues();
-                    parameters.put("index_name", index.indexName);
-                    parameters.put("index_type", index.indexType.toString());
-                    parameters.put("index_settings", settingsAsJSON(index));
-                    parameters.put("field_name", fieldName.field);
-                    parameters.put("last_sequence", 0);
-                    long rowId = database.insert(QueryImpl.INDEX_METADATA_TABLE_NAME,
-                                                 parameters);
-                    if (rowId < 0) {
-                        throw new QueryException("Error inserting index metadata");
-                    }
-                }
-
-                // Create SQLite data structures to support the index
-                // For JSON index type create a SQLite table and a SQLite index
-                // For TEXT index type create a SQLite virtual table
-                List<String> columnList = new ArrayList<String>();
-                for (FieldSort field: fieldNamesList) {
-                    columnList.add("\"" + field.field + "\"");
-                }
-
-                List<String> statements = new ArrayList<String>();
-                if (index.indexType == IndexType.TEXT) {
-                    String settings = String.format("tokenize=%s", index.tokenize);
-                    statements.add(createVirtualTableStatementForIndex(index.indexName,
-                            columnList,
-                            Collections.singletonList(settings)));
-                } else {
-                    statements.add(createIndexTableStatementForIndex(index.indexName, columnList));
-                    statements.add(createIndexIndexStatementForIndex(index.indexName, columnList));
-                }
-                for (String statement : statements) {
-                    try {
-                        database.execSQL(statement);
-                    } catch (SQLException e) {
-                        String msg = String.format("Index creation error occurred (%s):",statement);
-                        throw new QueryException(msg, e);
-                    }
-                }
-                return null;
-            }
-        });
+        Future<Void> result = queue.submitTransaction(new CreateIndexCallable(fieldNamesList, index));
 
         // Update the new index if it's been created
         try {
@@ -240,19 +188,6 @@ class IndexCreator {
                 queue);
 
         return index.indexName;
-    }
-
-    /**
-     * Helper to convert the index settings to a JSON string
-     *
-     * @return the JSON representation of the index settings
-     */
-    protected static String settingsAsJSON(Index i) {
-        // this is a trivial enough operation that we don't need a JSON serializer
-        if (i.tokenize == null) {
-            return "{}";
-        }
-        return "{\"tokenize\":\""+i.tokenize+"\"}";
     }
 
     /**
@@ -301,44 +236,7 @@ class IndexCreator {
         return false;
     }
 
-    private String createIndexTableStatementForIndex(String indexName, List<String> columns) {
-        String tableName = String.format(Locale.ENGLISH, "\"%s\"", QueryImpl.tableNameForIndex(indexName));
-        String cols = Misc.join(" NONE, ", columns);
 
-        return String.format("CREATE TABLE %s ( %s NONE )", tableName, cols);
-    }
-
-    private String createIndexIndexStatementForIndex(String indexName, List<String> columns) {
-        String tableName = QueryImpl.tableNameForIndex(indexName);
-        String sqlIndexName = tableName.concat("_index");
-        String cols = Misc.join(",", columns);
-
-        return String.format(Locale.ENGLISH, "CREATE INDEX \"%s\" ON \"%s\" ( %s )", sqlIndexName, tableName, cols);
-    }
-
-    /**
-     * This method generates the virtual table create SQL for the specified index.
-     * Note:  Any column that contains an '=' will cause the statement to fail
-     *        because it triggers SQLite to expect that a parameter/value is being passed in.
-     *
-     * @param indexName the index name to be used when creating the SQLite virtual table
-     * @param columns the columns in the table
-     * @param indexSettings the special settings to apply to the virtual table -
-     *                      (only 'tokenize' is current supported)
-     * @return the SQL to create the SQLite virtual table
-     */
-    private String createVirtualTableStatementForIndex(String indexName,
-                                                       List<String> columns,
-                                                       List<String> indexSettings) {
-        String tableName = String.format(Locale.ENGLISH, "\"%s\"", QueryImpl
-                .tableNameForIndex(indexName));
-        String cols = Misc.join(",", columns);
-        String settings = Misc.join(",", indexSettings);
-
-        return String.format("CREATE VIRTUAL TABLE %s USING FTS4 ( %s, %s )", tableName,
-                                                                              cols,
-                                                                              settings);
-    }
 
     /**
      * Iterate candidate indexNames generated from the indexNameRandom generator
