+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/sync/internal/query/callables/CreateIndexCallableTests.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright © 2017 IBM Corp. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *  except in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ *
+ */
+
+package com.cloudant.sync.internal.query.callables;
+
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+
+import com.cloudant.sync.query.FieldSort;
+import com.cloudant.sync.query.Index;
+import com.cloudant.sync.query.IndexType;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Created by rhys on 10/01/2017.
+ */
+public class CreateIndexCallableTests {
+
+    private List<FieldSort> fieldNames;
+    private String indexName;
+
+    @Before
+    public void setUp() {
+        fieldNames = Arrays.<FieldSort>asList(new FieldSort("name"), new FieldSort("age"));
+        indexName = "basic";
+
+    }
+
+    @Test
+    public void returnsIndexSettingsAsAString() {
+        Index index = new Index(fieldNames, indexName, IndexType.TEXT);
+        assertThat(CreateIndexCallable.settingsAsJSON(index), is("{\"tokenize\":\"simple\"}"));
+    }
+
+
+    @Test
+    public void returnsNullIndexSettingsAsAString() {
+        Index index = new Index(fieldNames, indexName);
+        assertThat(CreateIndexCallable.settingsAsJSON(index), is("{}"));
+    }
+}
